requirements.txt
accessible-output2==0.17
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
comtypes==1.4.16
cryptography==46.0.7
distro==1.9.0
google-auth==2.49.2
google-genai==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
keyboard==0.13.5
libloader==1.4.3
MouseInfo==0.1.3
packaging==26.1
pefile==2024.8.26
pillow==12.2.0
platform_utils==1.6.2
platformdirs==4.9.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.13.2
pydantic_core==2.46.2
PyGetWindow==0.0.9
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
PyMsgBox==2.0.1
pynput==1.8.1
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.2.2
pyttsx3==2.99
pytweening==1.2.0
pywin32==311
pywin32-ctypes==0.2.3
requests==2.33.1
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
wxPython==4.2.5
